Check for tmux/vim files before moving

+++ init.sh
@@ -9,8 +9,12 @@ git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 mkdir -p ~/.vim/autoload ~/.vim/bundle && \
 curl -LSso ~/.vim/autoload/pathogen.vim https://tpo.pe/pathogen.vim
 
-mv ~/.vimrc ~/.vimrc_bak
-mv ~/.tmux.conf ~/.tmux.conf_bak
+if [ -f ~/.vimrc ]; then
+    mv ~/.vimrc ~/.vimrc_bak
+fi
 
-cp ./vimrc ~/vimrc
+if [ -f ~/.tmux.conf ]; then
+    mv ~/.tmux.conf ~/.tmux.conf_bak
+fi
+cp ./.vimrc ~/.vimrc
 cp ./.tmux.conf ~/.tmux.conf_bak
